python/requirements.txt
openai==1.70.0
python-pptx==1.0.2
Pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3 